requirements.txt
certifi==2017.4.17
chardet==3.0.3
coverage==4.4.1
coveralls==1.1
docopt==0.6.2
idna==2.5
nomenclate==1.3.0
nose==1.3.7
python-dateutil==2.6.0
PyYAML==3.12
Qt.py==0.6.9
requests==2.17.3
six==1.10.0
urllib3==1.21.1
simplejson==3.10.0